Clarify Settings Ownership and Rename Storage Adapter
- Move Settings and defaultSettings into src/shared/settings.ts
- Rename googleStorage.ts to settingsStorage.ts for Chrome storage only
- Remove the settings type/default ownership cycle
- Update imports, docs, and the 1.2.4 changelog
- Keep validation, tests, and production build green
