--- src/shared/utils/data.ts
@@ -1,13 +1,0 @@
-import { SettingsType } from "@src/lib/utilities/googleStorage";
-
-export const defaultSettings: SettingsType = {
-    messageMaxWidthStyle: "95",
-    messageColorUserStyle: "#0084FF",
-    messageColorNonUserStyle: "#333333",
-    messagePaddingStyle: "10",
-    messageBorderRadiusStyle: "5",
-    inputBoxMaxWidthStyle: "94",
-    textColorUserStyle: "#FFFFFF",
-    textColorNonUserStyle: "#FFFFFF",
-    messageButtonsVisibilityStyle: true,
-};
